[Kernels] Allocate the paged KVCache based on available memory, not fixed batch_size/seq_len
We need to change the way we allocate memory for the paged kv cache.
This just extends the existing memory estimates to just take the
available memory off the top after the model loads, and fill that with
KVCache entries. We'll need to follow up on this to ensure we don't
over-subscribe and fall over.

MODULAR_ORIG_COMMIT_REV_ID: 5c7e15cb13966a9f9d1a0551b38a010a03b7f4ad