[Examples] [Graph API] Add Mandelbrot custom kernel example
Expanding upon the initial custom operations example, this adds a
standalone example that shows how to calculate the number of iterations
to escape in the Mandelbrot set. It illustrates how to pass a scalar
argument to a custom Mojo kernel, and how to specify a different
datatype for output of the kernel than the input.

MODULAR_ORIG_COMMIT_REV_ID: 080756f58a022a582b243b6c21a6db98941847f2