[Pipelines] Remove HuggingFaceFile and get_llama_huggingface_file utility
This removes deprecate methods which explictly use the HuggingFaceFile
utility. This only use in the codebase in which Llama weights should be
explcitly identified is in the GUI in which it is explictly downloading
the files using streamlit utilities. As such, I've moved the needed
logic into those files specifically, and removed shared utilities to
reduce the surface area to manage.

MODULAR_ORIG_COMMIT_REV_ID: 932286b6993708a39e382e8bfd0cbc62623a87b7