[KVCache] Extend paged attention KV cache manager with prefix caching (#53190)
During `manager.fetch`, we pass in a token list or prompt. Prior to this
change, these prompt tokens are assumed to not have a KV cache entry.
With prefix caching, we will look each token up in the prefix cache and
add the corresponding blocks to the sequence's set of committed blocks.
Now, we need to shorten the prompt to only include the tokens that were
not already found in the prefix cache.

One edge case is when all prompt tokens are found in the prefix cache.
In this case, we could naively return an empty prompt as each token is
already cached. However, this does not work as the model expects a
prompt of length at least 1. Thus, our prefix cache lookup attempts to
match on `prompt[:-1]` instead of the entire prompt.

Currently the implementation of `fetch` mutates the input
`seq_ids_and_prompt` to shorten the prompt. This is not a particularly
elegant api. We will come up with a better api in the future.

This PR does not support evictions yet. Once a block is committed and
inserted into the trie, it will not be released even after the original
sequence is released. This will allow future sequences to reuse the
computed KV entries, but also means that we never free the blocks even
when we are running out of room.

This PR only supports prefix caching when paged KVCache's page size is
1. This restriction will be lifted in the future.

Run it:
```
br //SDK/public/max-repo/pipelines/python:pipelines -- llama3 --quantization-encoding bfloat16 \
    --use-gpu --prompt "The meaning of life is" \
    --cache-strategy paged --kv-cache-page-size 1 --enable-prefix-caching
```

In the above example, we reduce the KV cache usage by half when we run
the same prompt twice (warmup + generation).

Previously for the first pass, we use blocks 0-511. For the second pass,
we use blocks 512-1023.

With prefix caching, we only use blocks 0-511 for both passes.

Fixes AIPIPE-217

MODULAR_ORIG_COMMIT_REV_ID: bdaa058659930995d331674ce16e758b41b188f2