[SDK] Add trim_prompt to TextAndVisionContext
Add `trim_prompt` to `TextAndVisionContext`.
This was added in #53190 but `TextAndVisionContext` was missed.
`TextAndVisionContext` should probably have a `TextContext` field or
subclass that or a common parent, but leave that for a follow up for
now.

MODULAR_ORIG_COMMIT_REV_ID: 22ab2020bdba01edb2e0fa524507eff9c69f8b0e