[pipelines] Pixtral works on any one image. Uses dynamic dims.
Currently Pixtral hardcodes the image size because it was generating the
attention_mask for vision encoder in numpy. The 2D diagonal mask
generation requires casting the image shapes to ints which is only
available in the graph API for Static dims.

For now, we will generate the attention mask and pass it as input to the
graph until we eventually do it in kernels.

Tested by prompting Pixtral with images of different sizes.

MODULAR_ORIG_COMMIT_REV_ID: 2f1fc11aee3fb756b208a09b1ca75e1bdbd64b8d