[Pipelines] Minor fixes to gguf encoding detection and logging (#53900)
A couple of fixes for things I ran into when loading from a local GGUF
([to reproduce this
[Internal link]

Was trying to run `bazel run
//SDK/public/max-repo/pipelines/python:pipelines -- serve --weight-path
./Llama-3.2-3B-Instruct-Q4_K_M.gguf --huggingface-repo-id
meta-llama/Llama-3.2-3B-Instruct` and the model kept trying to download
weights even though the `--weight-path` is passed in. Later noticed that
it was because the HUGGINGFACE engine was being used instead of the MAX
engine.

1. Added logging for why the engine was changed (there was already a
message defined that was never logged)
2. Made sure to lowercase the name when detecting GGUF encoding

MODULAR_ORIG_COMMIT_REV_ID: dd03bc5f5b3cb6f095ef60b5dfbf09eb4903364d