[Docs] [Examples] Update examples README to replace mention of Mojo pipelines with custom ops
The README for the MAX examples was referencing the previously-removed
Mojo pipelines and was lacking a section on the new custom ops examples.
The README has been updated to reflect this, and an additional
description was added for the new top-K sampler example.

MODULAR_ORIG_COMMIT_REV_ID: 3062bf4ae1460dc575a227f742b02854a246884d