[Pipelines] Implement the First Draft of WhisperEncoder
Added the implementation and the graph for the Encoder.

Currently this implementation is missing the following layers:

1) Conv1D with Bias. It turns out Conv1D has bias here. Not sure if bias
is implemented in our conv kernels. If so, will add it to both Conv1D
and Conv2D. Otherwise, will have to add it to the kernel.

2) AttentionWithoutMask with Bias. For Whispwe, q, v, and out
projections in attention have a bias vector. Current kernel doesn't take
it a bias vector.

Its not tested yet. Will be tested after these layers are added.

MODULAR_ORIG_COMMIT_REV_ID: 20b2c2099d3527018dd49ff728a029ac47dab6f6