[docs] Add Sphinx config files for pipeline.nn package (#55252)
Generate docs for the `max.pipelines.nn` package (all modules except
those in the `hooks` sub-package).

This also adds a new bash script to assist with this kind of task. It
will generate these .rst files based on a given directory of .py files.
It takes care of the tedious work but there’s still a little bit of
manual work to be done (although we could probably upgrade the script to
do more of it). But we can’t entirely rely on this sort of program; it’s
just to initialize the files and we must still return to add more
package/module summaries where appropriate in the RST files.

For this PR, I’m just going for the MVP version of docs. No linting of
docstrings and no new docstrings.

Fixes DOCS-703

MODULAR_ORIG_COMMIT_REV_ID: 09bedd574517b29dcb8ee0bc8d4d2da08ccc8401