[KVCache] Improve error messages in cases where we run out of KV cache blocks (Try 2) (#55342)
Recently we saw some regressions due to the KV cache running out of
blocks. e.g:
[Internal link]

In the above case, the KV manager actually allocated a total of 0 blocks
during init.

This would have been trivially obvious if the error message when we oom
printed the total number of blocks in the cache.

We also provide more context in many situations and catch edge cases
earlier.

This is same as #55257 which was reverted in #55288. In this PR we dont
check if we can support one request up to the max sequence length.

MODULAR_ORIG_COMMIT_REV_ID: eab7c7a831e103db8ca0fea40a1a84b81f22b9ac