[Pipelines] Clean up generation even if interrupted (#55291)
Previously, we'd only record end-generation metrics and mark the context
as complete if we had completed a generation on an EOS. If the user
interrupted the generation, for example, we'd say the tokens-per-second
was "N/A". I think it's more useful to clean up and emit metrics on the
tokens we generated even if we did not encounter an EOS.

Related to AITLIB-132.

MODULAR_ORIG_COMMIT_REV_ID: 36622146e84c78378d14ee3a04200760f5edf335