[Pipelines] Add `clip_qkv` to attention layers for Olmo models (#56666)
This was accidentally excluded from the Llama3 refactor in #56047

Tested with `allenai/OLMo-1B-0724-hf`, but the clip range is set to 8
while the attention weights are in the 0-0.05 range, so the outputs
don't actually change. I was able to verify that the weights are clamped
by manually setting the clip_qkv to 0.01.

Fixes MODELS-445

MODULAR_ORIG_COMMIT_REV_ID: f7ef969064b5116f9b745887090a5274144af2b2