[stdlib] Upgrade `Dict._find_index` to use `Pointer`.
Hello,
this pr upgrade the method to use `Pointer` to iterate `self._entries`.
(removing a `__copyinit__` on `K` and `V`)

The method is used by almost all methods of `Dict`,
that should be a great speedup with `K`,`V` that allocates.

(PR is a split of a previous PR :thumbsup)

Signed-off-by: rd4com <144297616+rd4com@users.noreply.github.com>