[stdlib] Add a parameter for inserting `llvm.assume` in `debug_assert`
In principle, `debug_assert` enforces an invariant that must hold at runtime. This information can also be valuable to the compiler for optimisation. Therefore, we added a `llvm_assume` parameter to `debug_assert` so it can emit an `llvm.assume` even when error reporting is disabled. We don't enable it by default, however, because in cases like using `select` over a "soft UB" in `prev_power_of_two`, emitting an `llvm.assume` could turn that soft UB into a hard one.

Signed-off-by: Yiwu Chen <210at85@gmail.com>